+++ core/extensions/impl/storage_extension.hpp
@@ -3,154 +3,131 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef KAGOME_STORAGE_EXTENSION_HPP
-#define KAGOME_STORAGE_EXTENSION_HPP
+#ifndef KAGOME_EXTENSIONS_STORAGE_EXTENSION_IMPL_HPP
+#define KAGOME_EXTENSIONS_STORAGE_EXTENSION_IMPL_HPP
 
+#include <cstdint>
+
+#include "common/logger.hpp"
+#include "extensions/impl/storage_extension.hpp"
+#include "runtime/trie_storage_provider.hpp"
 #include "runtime/wasm_memory.hpp"
+#include "storage/changes_trie/changes_tracker.hpp"
 
 namespace kagome::extensions {
   /**
    * Implements extension functions related to storage
    */
-  class StorageExtension {
+  class StorageExtensionImpl : public StorageExtension {
    public:
-    virtual ~StorageExtension() = default;
+    StorageExtensionImpl(
+        std::shared_ptr<runtime::TrieStorageProvider> storage_provider,
+        std::shared_ptr<runtime::WasmMemory> memory,
+        std::shared_ptr<storage::changes_trie::ChangesTracker> changes_tracker);
 
     // -------------------------Data storage--------------------------
-    /**
-     * @see Extension::ext_clear_prefix
-     */
-    virtual void ext_clear_prefix(runtime::WasmPointer prefix_data,
-                                  runtime::WasmSize prefix_length) = 0;
 
-    /**
-     * @see Extension::ext_clear_storage
-     */
-    virtual void ext_clear_storage(runtime::WasmPointer key_data,
-                                   runtime::WasmSize key_length) = 0;
+    void ext_clear_prefix(runtime::WasmPointer prefix_data,
+                          runtime::WasmSize prefix_length) override;
 
-    /**
-     * @see Extension::ext_exists_storage
-     */
-    virtual runtime::WasmSize ext_exists_storage(
-        runtime::WasmPointer key_data, runtime::WasmSize key_length) const = 0;
+    void ext_clear_storage(runtime::WasmPointer key_data,
+                           runtime::WasmSize key_length) override;
 
-    /**
-     * @see Extension::ext_get_allocated_storage
-     */
-    virtual runtime::WasmPointer ext_get_allocated_storage(
+    runtime::WasmSize ext_exists_storage(
+        runtime::WasmPointer key_data,
+        runtime::WasmSize key_length) const override;
+
+    runtime::WasmPointer ext_get_allocated_storage(
         runtime::WasmPointer key_data,
         runtime::WasmSize key_length,
-        runtime::WasmPointer len_ptr) = 0;
+        runtime::WasmPointer len_ptr) override;
 
-    /**
-     * @see Extension::ext_get_storage_into
-     */
-    virtual runtime::WasmSize ext_get_storage_into(
+    runtime::WasmSize ext_get_storage_into(
         runtime::WasmPointer key_data,
         runtime::WasmSize key_length,
         runtime::WasmPointer value_data,
         runtime::WasmSize value_length,
-        runtime::WasmSize value_offset) = 0;
+        runtime::WasmSize value_offset) override;
 
-    /**
-     * @see Extension::ext_storage_read_version_1
-     */
-    virtual runtime::WasmSpan ext_storage_read_version_1(
+    runtime::WasmSpan ext_storage_read_version_1(
         runtime::WasmSpan key,
         runtime::WasmSpan value_out,
-        runtime::WasmOffset offset) = 0;
+        runtime::WasmOffset offset) override;
 
-    /**
-     * @see Extension::ext_set_storage
-     */
-    virtual void ext_set_storage(runtime::WasmPointer key_data,
-                                 runtime::WasmSize key_length,
-                                 runtime::WasmPointer value_data,
-                                 runtime::WasmSize value_length) = 0;
+    void ext_set_storage(runtime::WasmPointer key_data,
+                         runtime::WasmSize key_length,
+                         runtime::WasmPointer value_data,
+                         runtime::WasmSize value_length) override;
 
     // -------------------------Trie operations--------------------------
 
-    /**
-     * @see Extension::ext_blake2_256_enumerated_trie_root
-     */
-    virtual void ext_blake2_256_enumerated_trie_root(
+    void ext_blake2_256_enumerated_trie_root(
         runtime::WasmPointer values_data,
         runtime::WasmPointer lengths_data,
         runtime::WasmSize values_num,
-        runtime::WasmPointer result) = 0;
+        runtime::WasmPointer result) override;
 
-    /**
-     * @see Extension::ext_storage_changes_root
-     */
-    virtual runtime::WasmPointer ext_storage_changes_root(
-        runtime::WasmPointer parent_hash, runtime::WasmPointer result) = 0;
+    runtime::WasmPointer ext_storage_changes_root(
+        runtime::WasmPointer parent_hash, runtime::WasmPointer result) override;
 
-    /**
-     * @see Extension::ext_storage_root
-     */
-    virtual void ext_storage_root(runtime::WasmPointer result) const = 0;
+    void ext_storage_root(runtime::WasmPointer result) const override;
 
     // ------------------------ VERSION 1 ------------------------
 
-    /**
-     * @see Extension::ext_storage_set_version_1
-     */
-    virtual void ext_storage_set_version_1(runtime::WasmSpan key,
-                                           runtime::WasmSpan value) = 0;
+    void ext_storage_set_version_1(runtime::WasmSpan key,
+                                   runtime::WasmSpan value) override;
 
-    /**
-     * @see Extension::ext_storage_get_version_1
-     */
-    virtual runtime::WasmSpan ext_storage_get_version_1(
-        runtime::WasmSpan key) = 0;
+    runtime::WasmSpan ext_storage_get_version_1(runtime::WasmSpan key) override;
 
-    /**
-     * @see Extension::ext_storage_clear_version_1
-     */
-    virtual void ext_storage_clear_version_1(runtime::WasmSpan key_data) = 0;
+    void ext_storage_clear_version_1(runtime::WasmSpan key_data) override;
 
-    /**
-     * @see Extension::ext_storage_exists_version_1
-     */
-    virtual runtime::WasmSize ext_storage_exists_version_1(
-        runtime::WasmSpan key_data) const = 0;
+    runtime::WasmSize ext_storage_exists_version_1(
+        runtime::WasmSpan key_data) const override;
 
-    /**
-     * @see Extension::ext_storage_clear_prefix_version_1
-     */
-    virtual void ext_storage_clear_prefix_version_1(
-        runtime::WasmSpan prefix) = 0;
+    void ext_storage_clear_prefix_version_1(runtime::WasmSpan prefix) override;
 
-    /**
-     * @see Extension::ext_storage_root_version_1
-     */
-    virtual runtime::WasmPointer ext_storage_root_version_1() const = 0;
+    runtime::WasmPointer ext_storage_root_version_1() const override;
 
-    /**
-     * @see Extension::ext_storage_changes_root_version_1
-     */
-    virtual runtime::WasmPointer ext_storage_changes_root_version_1(
-        runtime::WasmSpan parent_hash) = 0;
+    runtime::WasmPointer ext_storage_changes_root_version_1(
+        runtime::WasmSpan parent_hash) override;
 
-    /**
-     * @see Extension::ext_storage_next_key
-     */
-    virtual runtime::WasmSpan ext_storage_next_key_version_1(
-        runtime::WasmSpan key) const = 0;
+    runtime::WasmSpan ext_storage_next_key_version_1(
+        runtime::WasmSpan key) const override;
 
-    /**
-     * @see Extension::ext_trie_blake2_256_root_version_1
-     */
-    virtual runtime::WasmPointer ext_trie_blake2_256_root_version_1(
-        runtime::WasmSpan values_data) = 0;
+    runtime::WasmPointer ext_trie_blake2_256_root_version_1(
+        runtime::WasmSpan values_data) override;
 
+    runtime::WasmPointer ext_trie_blake2_256_ordered_root_version_1(
+        runtime::WasmSpan values_data) override;
+
+   private:
     /**
-     * @see Extension::ext_trie_blake2_256_ordered_root_version_1
+     * Find the value by given key and the return the part of it starting from
+     * given offset
+     *
+     * @param key Buffer representation of the key
+     * @param offset SizeType pointing to the beginning of the value
+     * @param max_length SizeType defining the maximum possible length of the
+     * returned result
+     * @return result containing Buffer with the part of the value, or error in
+     * case value by give key does not exist
      */
-    virtual runtime::WasmPointer ext_trie_blake2_256_ordered_root_version_1(
-        runtime::WasmSpan values_data) = 0;
+    outcome::result<common::Buffer> get(const common::Buffer &key,
+                                        runtime::WasmSize offset,
+                                        runtime::WasmSize max_length) const;
+
+    outcome::result<common::Buffer> get(const common::Buffer &key) const;
+
+    outcome::result<boost::optional<common::Buffer>> getStorageNextKey(
+        const common::Buffer &key) const;
+
+    std::shared_ptr<runtime::TrieStorageProvider> storage_provider_;
+    std::shared_ptr<runtime::WasmMemory> memory_;
+    std::shared_ptr<storage::changes_trie::ChangesTracker> changes_tracker_;
+    common::Logger logger_;
+
+    constexpr static auto kDefaultLoggerTag = "WASM Runtime [StorageExtension]";
   };
 }  // namespace kagome::extensions
 
-#endif  // KAGOME_STORAGE_EXTENSION_HPP=
+#endif  // KAGOME_EXTENSIONS_STORAGE_EXTENSION_IMPL_HPP
